Catch tts.getVoices error
On Samsung Galaxy S4 there is an unexpected error that happens sometimes when users are using the Samsung default voice and tts.getVoices is called. This catches the error and notes the problem, falling back to the old approach to getting a list of locales since the voice list is not available.